primarily work on making ncurses/libtinfo easier to link to
by effectively dropping the ABI 5 ncurses as a build target
and not building libtinfo (it is integrated into ncurses as it should be)
in the ABI 6 version.

Also, new nano to test new ncurses, cmake that was easier
to build without the -ltinfo hassle. Still need to work
on newer cmake at some point.